fix(application): 패키지 오타 수정
작업 내용
- validator 패키지명 오타 수정
