+++ src/main/java/lotto/validator/LogicValidator.java
@@ -1,9 +1,10 @@
-package lotto.vaildator;
+package lotto.validator;
+
+import lotto.constant.ExceptionMessage;
 
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import lotto.constant.ExceptionMessage;
 
 public class LogicValidator {
