+++ src/test/java/lotto/vaildator/InputValidatorTest.java
@@ -1,14 +1,16 @@
 package lotto.vaildator;
 
-import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-
-import java.util.List;
 import lotto.constant.ExceptionMessage;
 import lotto.domain.Lotto;
+import lotto.validator.InputValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 class InputValidatorTest {
