+++ src/test/java/lotto/vaildator/LogicValidatorTest.java
@@ -1,13 +1,15 @@
 package lotto.vaildator;
 
-import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-
-import java.util.List;
 import lotto.constant.ExceptionMessage;
+import lotto.validator.LogicValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 public class LogicValidatorTest {
